fix: remove walk through walls, and minor fixes

+++ include/item.h
@@ -13,7 +13,7 @@ struct Item
     u16 itemId;
     u16 price;
     u8 holdEffect;
-    u16 holdEffectParam;
+    u8 holdEffectParam;
     const u8 *description;
     u8 importance;
     bool8 registrability; // unused
@@ -68,7 +68,7 @@ bool8 RemovePyramidBagItem(u16 itemId, u16 count);
 const u8 *ItemId_GetName(u16 itemId);
 u16 ItemId_GetPrice(u16 itemId);
 u8 ItemId_GetHoldEffect(u16 itemId);
-u16 ItemId_GetHoldEffectParam(u16 itemId);
+u8 ItemId_GetHoldEffectParam(u16 itemId);
 const u8 *ItemId_GetDescription(u16 itemId);
 u8 ItemId_GetImportance(u16 itemId);
 u8 ItemId_GetPocket(u16 itemId);
